fix: 모델수정, cham&kamis 데이터 연결

+++ Projects/Soonsu/View/MyMenu/MyMenuView/Component/MenuRowView.swift
@@ -76,7 +76,7 @@ struct MenuRowView: View {
     }
     
     /// 총 원가(Int)
-    private var totalCost: Int {
+    private var totalCost: Double {
         matchedEntities.reduce(0) { $0 + $1.unitPrice }
     }
